refactor cheapgm
